# Conflicts:
#	app/src/main/java/com/mnassa/screen/events/details/participants/EventParticipantsRVAdapter.kt
#	app/src/main/java/com/mnassa/screen/posts/need/sharing/SharingOptionsController.kt
#	app/src/main/res/controller/layout/controller_connections.xml
#	app/src/main/res/values/dictionary.xml

+++ app/src/main/java/com/mnassa/screen/comments/rewarding/RewardingViewModel.kt
@@ -9,5 +9,5 @@ import kotlinx.coroutines.experimental.channels.BroadcastChannel
  * Date: 4/30/2018
  */
 interface RewardingViewModel : MnassaViewModel {
-    val defaultRewardChannel: BroadcastChannel<Int>
+    val defaultRewardChannel: BroadcastChannel<Long>
 }
